+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+  - Deployment log - Bugfix deployment log need job get-context
   - Nuvlabox status - Next-heartbeat set 2 * refresh-interval + 10s
   - Deployment - Bugfix always recreate deployment creds at start
   - Nuvlabox - Update operation support payload